+++ ts/features/itwallet/presentation/proximity/machine/actors.ts
@@ -1,5 +1,4 @@
 import { ISO18013_5 } from "@pagopa/io-react-native-iso18013";
-import BluetoothStateManager from "react-native-bluetooth-state-manager";
 import { fromCallback, fromPromise } from "xstate";
 import { assert } from "../../../../../utils/assert";
 import { Env } from "../../../common/utils/environment";
@@ -17,7 +16,10 @@ import {
   PROXIMITY_PERMISSIONS_TO_CHECK
 } from "../utils";
 import type { EventsPayload } from "../utils/itwProximityTypeUtils";
-import { checkBluetoothPermissions } from "../utils/permissions";
+import {
+  checkBluetoothPermissions,
+  checkBluetoothActivation
+} from "../utils/ble";
 import { Context } from "./context";
 import { ProximityEvents } from "./events";
 
@@ -41,14 +43,13 @@ export type SendDocumentsActorOutput = Awaited<
 >;
 
 export const createProximityActorsImplementation = (env: Env) => {
-  const checkBluetoohPermissions = fromPromise<boolean>(
+  const checkBluetoothPermissionsActor = fromPromise<boolean>(
     checkBluetoothPermissions
   );
 
-  const checkBluetoothActivation = fromPromise<boolean>(async () => {
-    const bluetoothState = await BluetoothStateManager.getState();
-    return bluetoothState === "PoweredOn";
-  });
+  const checkBluetoothActivationActor = fromPromise<boolean>(
+    checkBluetoothActivation
+  );
 
   const checkBluetoothIsActive = fromPromise<boolean, void>(
     isBluetoothPoweredOn
@@ -201,8 +202,8 @@ export const createProximityActorsImplementation = (env: Env) => {
   );
 
   return {
-    checkBluetoohPermissions,
-    checkBluetoothActivation,
+    checkBluetoothPermissions: checkBluetoothPermissionsActor,
+    checkBluetoothActivation: checkBluetoothActivationActor,
     startEngagement,
     proximityCommunicationLogic,
     closeProximityFlow,
