+++ ts/features/itwallet/presentation/proximity/navigation/ItwProximityParamsList.ts
@@ -3,7 +3,7 @@ import { ITW_PROXIMITY_ROUTES } from "./routes";
 
 export type ItwProximityParamsList = {
   [ITW_PROXIMITY_ROUTES.QR_CODE]: ItwProximityQrCodeScreenNavigationParams;
-  [ITW_PROXIMITY_ROUTES.DEVICE_PERMISSIONS]: undefined;
+  [ITW_PROXIMITY_ROUTES.BLUETOOTH_PERMISSIONS]: undefined;
   [ITW_PROXIMITY_ROUTES.BLUETOOTH_ACTIVATION]: undefined;
   [ITW_PROXIMITY_ROUTES.CLAIMS_DISCLOSURE]: undefined;
   [ITW_PROXIMITY_ROUTES.SEND_DOCUMENTS_RESPONSE]: undefined;
