+++ ts/features/itwallet/presentation/proximity/navigation/routes.ts
@@ -1,7 +1,7 @@
 export const ITW_PROXIMITY_ROUTES = {
   MAIN: "ITW_PROXIMITY_MAIN",
   QR_CODE: "ITW_PROXIMITY_QR_CODE",
-  DEVICE_PERMISSIONS: "ITW_PROXIMITY_DEVICE_PERMISSIONS",
+  BLUETOOTH_PERMISSIONS: "ITW_PROXIMITY_BLUETOOTH_PERMISSIONS",
   BLUETOOTH_ACTIVATION: "ITW_PROXIMITY_BLUETOOTH_ACTIVATION",
   CLAIMS_DISCLOSURE: "ITW_PROXIMITY_CLAIMS_DISCLOSURE",
   SEND_DOCUMENTS_RESPONSE: "ITW_PROXIMITY_SEND_DOCUMENTS_RESPONSE",
