+++ ts/features/itwallet/presentation/proximity/utils/ble.ts
@@ -1,4 +1,5 @@
-import { Platform } from "react-native";
+import { Linking, Platform } from "react-native";
+import AndroidOpenSettings from "react-native-android-open-settings";
 import {
   checkMultiple,
   Permission,
@@ -6,6 +7,7 @@ import {
   requestMultiple,
   RESULTS
 } from "react-native-permissions";
+import BluetoothStateManager from "react-native-bluetooth-state-manager";
 
 const BLUETOOTH_PERMISSIONS: Array<Permission> =
   Platform.OS === "android"
@@ -45,5 +47,28 @@ export const checkBluetoothPermissions = async () => {
     // Verify if all requested permissions are granted
     return allPermissionsGranted;
   }
+
   return true;
+};
+
+/**
+ * Checks if Bluetooth is currently activated on the device.
+ * @returns A promise that resolves to true if Bluetooth is powered on, or false otherwise.
+ */
+export const checkBluetoothActivation = async () => {
+  const bluetoothState = await BluetoothStateManager.getState();
+  return bluetoothState === "PoweredOn";
+};
+
+/**
+ * Opens the Bluetooth settings page on the device. On iOS, it attempts to open
+ * the Bluetooth settings directly, while on Android it opens the Bluetooth
+ * settings screen.
+ */
+export const openBluetoothPreferences = () => {
+  if (Platform.OS === "ios") {
+    Linking.openURL("App-Prefs:Bluetooth").catch(() => null);
+  } else {
+    AndroidOpenSettings.bluetoothSettings();
+  }
 };
